+++ src/main/java/oshi/hardware/platform/mac/MacHardwareAbstractionLayer.java
@@ -26,6 +26,7 @@ import oshi.hardware.Display;
 import oshi.hardware.GlobalMemory;
 import oshi.hardware.HardwareAbstractionLayer;
 import oshi.hardware.PowerSource;
+import oshi.hardware.Sensors;
 import oshi.software.os.OSFileStore;
 import oshi.software.os.mac.MacFileSystem;
 
@@ -40,6 +41,8 @@ public class MacHardwareAbstractionLayer implements HardwareAbstractionLayer {
     private GlobalMemory _memory;
 
     private PowerSource[] _powerSources;
+
+    private Sensors sensors;
 
     private JsonBuilderFactory jsonFactory = Json.createBuilderFactory(null);
 
@@ -88,6 +91,14 @@ public class MacHardwareAbstractionLayer implements HardwareAbstractionLayer {
     }
 
     @Override
+    public Sensors getSensors() {
+        if (this.sensors == null) {
+            this.sensors = new MacSensors();
+        }
+        return this.sensors;
+    }
+
+    @Override
     public JsonObject toJSON() {
         JsonArrayBuilder powerSourceArrayBuilder = jsonFactory.createArrayBuilder();
         for (PowerSource powerSource : getPowerSources()) {
@@ -103,6 +114,7 @@ public class MacHardwareAbstractionLayer implements HardwareAbstractionLayer {
         }
         return jsonFactory.createObjectBuilder().add("processor", getProcessor().toJSON())
                 .add("memory", getMemory().toJSON()).add("powerSources", powerSourceArrayBuilder.build())
-                .add("fileStores", fileStoreArrayBuilder.build()).add("displays", displayArrayBuilder.build()).build();
+                .add("fileStores", fileStoreArrayBuilder.build()).add("displays", displayArrayBuilder.build())
+                .add("sensors", getSensors().toJSON()).build();
     }
 }